Fixed pause menu buttons not reacting to mouse & optimized script
The pause menu buttons now work as intended, this was most likely due to a layering issue in the scene.

The PauseMenu script has been optimized so that there are less lines of code needed.

Added GameManager class to serve as a data class for now, this is so that we can store data of the game, such as the Player's gameobject so that we don't need to refer to the player as "FindObjectOfType<PlayerController>" in most scripts.