Changes made to Player, Projectile, GrabItem and Scaling Gun
-- Scaling gun
Projectile now spawns facing the player's direction. 
Scaling gun is now capable of setting Projectile's force rather that relying on a static value of a prefab.

-- Projectile
Projectile's scale is now set to its lifetime.
Projectile will no longer effect the GameObject which the player is holding.

-- Player
Player will not be able to rotate their view if they are holding a GameObject, and are holding the R button

-- GrabItem
Player is now able to rotate the GameObject they are holding using the mouse.

-- GameManager
Player can now be accessed from the GameManager - Player must be initialized to use the getter!