Fixed portals
Portals now work as intended, however, currently only the player is allowed to enter the portals.

The portals are currently capable of being entered both sides of the render plane, so until the portal has an asset with some kind of back wall for the render plane, or can be entered from only the front, then this will remain as a feature.